refactor(matrix): Combine benchmark test suites
