+++ packages/dds/matrix/src/test/time/sharedMatrix.spec.ts
@@ -2,621 +2,6 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
+import { runBenchmarkTestSuite } from "../matrixBenchmarks.js";
 
-import { strict as assert } from "node:assert";
-
-import {
-	benchmark,
-	BenchmarkType,
-	isInPerformanceTestingMode,
-	type BenchmarkTimer,
-	type BenchmarkTimingOptions,
-} from "@fluid-tools/benchmark";
-import type { Test } from "mocha";
-
-import { createTestMatrix, type MatrixBenchmarkOptions } from "../performanceTestUtilities.js";
-
-/**
- * Note: These benchmarks are designed to closely match the benchmarks in SharedTree.
- * If you modify or add tests here, consider updating the corresponding SharedTree benchmarks as well
- * to ensure consistency and comparability between the two implementations.
- */
-
-/**
- * {@link runBenchmark} configuration.
- */
-interface BenchmarkConfig extends BenchmarkTimingOptions, MatrixBenchmarkOptions {
-	readonly maxBenchmarkDurationSeconds: number;
-}
-
-/**
- * Runs a benchmark for measuring the execution time of operations on a SharedMatrix.
- */
-function runBenchmark({
-	title,
-	matrixSize,
-	initialCellValue,
-	beforeOperation,
-	operation,
-	afterOperation,
-	minBatchDurationSeconds = 0,
-	maxBenchmarkDurationSeconds,
-}: BenchmarkConfig): Test {
-	return benchmark({
-		type: BenchmarkType.Measurement,
-		title,
-		benchmarkFnCustom: async <T>(state: BenchmarkTimer<T>) => {
-			let duration: number;
-			do {
-				assert.equal(state.iterationsPerBatch, 1, "Expected exactly one iteration per batch");
-
-				// Create matrix
-				const { matrix, undoRedoStack, cleanUp } = createTestMatrix({
-					matrixSize,
-					initialCellValue,
-				});
-
-				beforeOperation?.(matrix, undoRedoStack);
-
-				// Operation
-				const before = state.timer.now();
-				operation(matrix, undoRedoStack);
-				const after = state.timer.now();
-
-				// Measure
-				duration = state.timer.toSeconds(before, after);
-
-				afterOperation?.(matrix, undoRedoStack);
-
-				// Cleanup
-				cleanUp();
-			} while (state.recordBatch(duration));
-		},
-		minBatchDurationSeconds,
-		maxBenchmarkDurationSeconds,
-	});
-}
-
-describe("SharedMatrix execution time", () => {
-	// The value to be set in the cells of the matrix.
-	const initialCellValue = "cellValue";
-
-	// The test matrix's size will be 10*10, 100*100.
-	// Matrix size 1000 benchmarks removed due to high overhead and unreliable results.
-	const matrixSizes = isInPerformanceTestingMode
-		? [10, 100]
-		: // When not measuring perf, use a single smaller data size so the tests run faster.
-			[10];
-
-	// The number of operations to perform on the matrix.
-	// Operation counts 1000 removed due to high overhead and unreliable results.
-	const operationCounts = isInPerformanceTestingMode
-		? [10, 100]
-		: // When not measuring perf, use a single smaller data size so the tests run faster.
-			[5];
-	let maxBenchmarkDurationSeconds: number;
-
-	for (const matrixSize of matrixSizes) {
-		maxBenchmarkDurationSeconds = matrixSize === 100 ? 10 : 5;
-		describe(`Size of ${matrixSize}*${matrixSize} SharedMatrix`, () => {
-			// Filter counts to ensure remove operation do not exceed matrixSize
-			const validRemoveCounts = operationCounts.filter((count) => count <= matrixSize);
-
-			// Insert-related tests that are not limited by matrixSize
-			for (const count of operationCounts) {
-				describe(`Column insertion`, () => {
-					const scenarioName = `Insert a single column in the middle of the table ${count} times`;
-					runBenchmark({
-						title: scenarioName,
-						matrixSize,
-						initialCellValue,
-						operation: (matrix) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-							}
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Undo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.undoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Redo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-							assert.equal(undoRedoStack.undoStackLength, 0);
-							assert.equal(undoRedoStack.redoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.redoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.redoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-				});
-
-				describe(`Row insertion`, () => {
-					const scenarioName = `Insert a single row in the middle of the table ${count} times`;
-					runBenchmark({
-						title: scenarioName,
-						matrixSize,
-						initialCellValue,
-						operation: (matrix) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Undo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.undoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Redo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-							assert.equal(undoRedoStack.undoStackLength, 0);
-							assert.equal(undoRedoStack.redoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.redoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.redoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-				});
-
-				describe(`Column and row insertion`, () => {
-					const scenarioName = `Insert a single row and a single column in the middle of the table ${count} times`;
-					runBenchmark({
-						title: scenarioName,
-						matrixSize,
-						initialCellValue,
-						operation: (matrix) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-							}
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Undo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, 2 * count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < 2 * count; i++) {
-								undoRedoStack.undoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.undoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Redo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-							}
-							for (let i = 0; i < 2 * count; i++) {
-								undoRedoStack.undoOperation();
-							}
-							assert.equal(undoRedoStack.undoStackLength, 0);
-							assert.equal(undoRedoStack.redoStackLength, 2 * count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < 2 * count; i++) {
-								undoRedoStack.redoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.redoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-				});
-			}
-
-			// Set/Remove-related tests that are limited by matrixSize
-			for (const count of validRemoveCounts) {
-				describe(`Column removal`, () => {
-					const scenarioName = `Remove the middle column ${count} times`;
-					runBenchmark({
-						title: scenarioName,
-						matrixSize,
-						initialCellValue,
-						operation: (matrix) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-							}
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Undo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.undoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Redo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-							assert.equal(undoRedoStack.undoStackLength, 0);
-							assert.equal(undoRedoStack.redoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.redoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.redoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-				});
-
-				describe(`Row removal`, () => {
-					const scenarioName = `Remove the middle row ${count} times`;
-					runBenchmark({
-						title: scenarioName,
-						matrixSize,
-						initialCellValue,
-						operation: (matrix) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Undo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.undoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Redo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-							assert.equal(undoRedoStack.undoStackLength, 0);
-							assert.equal(undoRedoStack.redoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.redoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.redoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-				});
-
-				describe(`Single column and row removal`, () => {
-					const scenarioName = `Remove the middle row and column ${count} times`;
-					runBenchmark({
-						title: scenarioName,
-						matrixSize,
-						initialCellValue,
-						operation: (matrix) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Undo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, 2 * count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < 2 * count; i++) {
-								undoRedoStack.undoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.undoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Redo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, 2 * count);
-							for (let i = 0; i < 2 * count; i++) {
-								undoRedoStack.undoOperation();
-							}
-							assert.equal(undoRedoStack.undoStackLength, 0);
-							assert.equal(undoRedoStack.redoStackLength, 2 * count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < 2 * count; i++) {
-								undoRedoStack.redoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.redoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-				});
-
-				describe(`Insert a row and a column and remove them right away`, () => {
-					const scenarioName = `Insert a row and a column and remove them right away ${count} times`;
-					runBenchmark({
-						title: scenarioName,
-						matrixSize,
-						initialCellValue,
-						operation: (matrix) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Undo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, 4 * count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < 4 * count; i++) {
-								undoRedoStack.undoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.undoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Redo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.insertCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.insertRows(Math.floor(matrix.rowCount / 2), 1);
-								matrix.removeCols(Math.floor(matrix.colCount / 2), 1);
-								matrix.removeRows(Math.floor(matrix.rowCount / 2), 1);
-							}
-							assert.equal(undoRedoStack.undoStackLength, 4 * count);
-							for (let i = 0; i < 4 * count; i++) {
-								undoRedoStack.undoOperation();
-							}
-							assert.equal(undoRedoStack.undoStackLength, 0);
-							assert.equal(undoRedoStack.redoStackLength, 4 * count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < 4 * count; i++) {
-								undoRedoStack.redoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.redoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-				});
-
-				describe(`Set cell values`, () => {
-					const scenarioName = `Set a 3-character string in ${count} cells`;
-					runBenchmark({
-						title: scenarioName,
-						matrixSize,
-						initialCellValue,
-						operation: (matrix) => {
-							for (let i = 0; i < count; i++) {
-								matrix.setCell(i, i, "abc");
-							}
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Undo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.setCell(i, i, "abc");
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.undoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-
-					runBenchmark({
-						title: `Redo: ${scenarioName}`,
-						matrixSize,
-						initialCellValue,
-						beforeOperation: (matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								matrix.setCell(i, i, "abc");
-							}
-							assert.equal(undoRedoStack.undoStackLength, count);
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.undoOperation();
-							}
-							assert.equal(undoRedoStack.undoStackLength, 0);
-							assert.equal(undoRedoStack.redoStackLength, count);
-						},
-						operation: (_matrix, undoRedoStack) => {
-							for (let i = 0; i < count; i++) {
-								undoRedoStack.redoOperation();
-							}
-						},
-						afterOperation: (_matrix, undoRedoStack) => {
-							assert.equal(undoRedoStack.redoStackLength, 0);
-						},
-						maxBenchmarkDurationSeconds,
-					});
-				});
-			}
-		});
-	}
-});
+runBenchmarkTestSuite("execution-time");
